refactor: 전역 사용 Enum 이동 및 domain 패키지 삭제
- domain/common/ShootingDurationOption -> global/common/enums/ShootingDurationOption 이동
- 최상위 domain 패키지 완전히 삭제하여 구조 정리
